4.3.8
If you don’t know where to start with your AI research journey, save this post now.

I’ve been neck-deep in AI for the last 18 months or so, and it’s been a LOT of experimenting and researching. I honestly feel like YouTube owes me money for all the content I have watched.

So, to save you some time, I've put together an AI starter pack that's got a bit of everything. This isn't the be-all and end-all, but it's a damn good place to kick things off…

YouTube Channels to Watch

↳ Matt Wolfe: Consistent, energetic, and all about creative AI. Your weekly digest of all things AI. It's like a cheat sheet, but fun.
https://lnkd.in/eK-AkHgp

↳ David Shapiro: Deep diver into societal impacts of AI. Macro views, big thoughts, and Star Trek uniforms.
https://lnkd.in/ehATRQ8W

↳ Matthew Berman: Jack of all trades on AI. Whether you want the big picture or a detailed analysis, Matthew's your man.
https://lnkd.in/e-dvHKRP

↳ AI Explained: The voice of reason in the AI buzz. Academic with a dash of opinion zero hype.
https://lnkd.in/e7ADf29A

Websites to Explore

↳ There's An AI For That: Find your AI match here, no matter what you need.
https://lnkd.in/e5RUkQvb

↳ AI Weirdness: Janelle Shane's playground of AI creativity. Get inspired, have a laugh.
https://lnkd.in/eHHGfiA6

↳ AI Dungeon: Storytelling meets gaming. Unique way to engage with AI and have fun while at it. https://aidungeon.com/

Especially Creative AI Tools

↳ Sora: Built by OpenAI, watch text prompts transform into full-fledged videos. Prepare to have your mind blown.
https://openai.com/sora

↳ HeyGen: An avatar that reads scripts in your voice?!
https://www.heygen.com/

↳ Adobe Photoshop Firefly: Retouching made easy with Generative Fill.
https://lnkd.in/ebPt8YP8

↳ LTX Studio: A powerful video content creation tool to watch out for.
https://ltx.studio/

↳ Pi AI: Not your typical creative tool, but a fascinating conversational, voice-activated platform for some AI fun.
https://pi.ai/talk

↳ Gemma.ai: Of course, our very own Gemma. Creativity and AI, all in one place:
https://lnkd.in/ePhKcX6p

Like I said, very much the tip of the AI iceberg, so feel free to drop in your recommendations in the comments! Don't hold back!